Added specs

+++ lib/danger/request_sources/gitlab.rb
@@ -69,15 +69,19 @@ module Danger
       end
 
       def mr_comments
+        # @raw_comments contains what we got back from the server.
+        # @comments contains Comment objects (that have less information)
         @comments ||= begin
           if supports_inline_comments
-            client.merge_request_discussions(ci_source.repo_slug, ci_source.pull_request_id)
+            @raw_comments = client.merge_request_discussions(ci_source.repo_slug, ci_source.pull_request_id)
               .auto_paginate
-              .flat_map(&:notes)
-              .map { |comment| Comment.from_gitlab_discussion(comment) }
+              .flat_map { |discussion| discussion.notes.map { |note| note.merge({"discussion_id" => discussion.id}) } }
+            @raw_comments
+              .map { |comment| Comment.from_gitlab(comment) }
           else
-            client.merge_request_comments(ci_source.repo_slug, ci_source.pull_request_id, per_page: 100)
+            @raw_comments = client.merge_request_comments(ci_source.repo_slug, ci_source.pull_request_id, per_page: 100)
               .auto_paginate
+            @raw_comments
               .map { |comment| Comment.from_gitlab(comment) }
           end
         end
@@ -135,14 +139,15 @@ module Danger
       def update_pull_request_with_inline_comments!(warnings: [], errors: [], messages: [], markdowns: [], danger_id: "danger", new_comment: false, remove_previous_comments: false)
         editable_comments = mr_comments.select { |comment| comment.generated_by_danger?(danger_id) }
 
-        should_create_new_comment = new_comment || editable_comments.empty? || remove_previous_comments
+        last_comment = editable_comments.last
+        should_create_new_comment = new_comment || last_comment.nil? || remove_previous_comments
 
-        if should_create_new_comment
-          previous_violations = {}
-        else
-          comment = editable_comments.first.body
-          previous_violations = parse_comment(comment)
-        end
+        previous_violations =
+          if should_create_new_comment
+            {}
+          else
+            parse_comment(last_comment.body)
+          end
 
         regular_violations = regular_violations_group(
           warnings: warnings,
@@ -162,6 +167,34 @@ module Danger
           danger_id: danger_id,
           previous_violations: previous_violations
         }.merge(inline_violations))
+
+        main_violations = merge_violations(
+          regular_violations, rest_inline_violations
+        )
+
+        main_violations_sum = main_violations.values.inject(:+)
+
+        if (previous_violations.empty? && main_violations_sum.empty?) || remove_previous_comments
+          # Just remove the comment, if there's nothing to say or --remove-previous-comments CLI was set.
+          delete_old_comments!(danger_id: danger_id)
+        end
+
+        # If there are still violations to show
+        if main_violations_sum.any?
+          body = generate_comment({
+            template: "gitlab",
+            danger_id: danger_id,
+            previous_violations: previous_violations
+          }.merge(main_violations))
+
+          comment_result =
+            if should_create_new_comment
+              client.create_merge_request_note(ci_source.repo_slug, ci_source.pull_request_id, body)
+            else
+              client.edit_merge_request_note(ci_source.repo_slug, ci_source.pull_request_id, last_comment.id, body)
+            end
+        end
+        
       end
 
       def update_pull_request_without_inline_comments!(warnings: [], errors: [], messages: [], markdowns: [], danger_id: "danger", new_comment: false, remove_previous_comments: false)
@@ -205,14 +238,21 @@ module Danger
       end
 
       def delete_old_comments!(except: nil, danger_id: "danger")
-        mr_comments.each do |comment|
+        @raw_comments.each do |raw_comment|
+
+          comment = Comment.from_gitlab(raw_comment)
           next unless comment.generated_by_danger?(danger_id)
           next if comment.id == except
-          client.delete_merge_request_comment(
-            ci_source.repo_slug,
-            ci_source.pull_request_id,
-            comment.id
-          )
+          next unless raw_comment.is_a?(Hash) && raw_comment["position"].nil?
+
+          begin
+            client.delete_merge_request_comment(
+              ci_source.repo_slug,
+              ci_source.pull_request_id,
+              comment.id
+            )
+          rescue
+          end
         end
       end
 
@@ -255,16 +295,18 @@ module Danger
         }
       end
 
-      def submit_inline_comments!(warnings: [], errors: [], messages: [], markdowns: [], previous_violations: [], danger_id: "danger")
-        # Avoid doing any fetchs if there's no inline comments
-        return {} if (warnings + errors + messages + markdowns).select(&:inline?).empty?
+      def merge_violations(*violation_groups)
+        violation_groups.inject({}) do |accumulator, group|
+          accumulator.merge(group) { |_, old, fresh| old + fresh }
+        end
+      end
 
-        # diff_lines = self.mr_diff.lines
+      def submit_inline_comments!(warnings: [], errors: [], messages: [], markdowns: [], previous_violations: [], danger_id: "danger")
         comments = client.merge_request_discussions(ci_source.repo_slug, ci_source.pull_request_id)
           .auto_paginate
           .flat_map { |discussion| discussion.notes.map { |note| note.merge({"discussion_id" => discussion.id}) } }
 
-        danger_comments = comments.select { |comment| Comment.from_gitlab_discussion(comment).generated_by_danger?(danger_id) }
+        danger_comments = comments.select { |comment| Comment.from_gitlab(comment).generated_by_danger?(danger_id) }
         non_danger_comments = comments - danger_comments
 
         diff_lines = []
@@ -329,17 +371,12 @@ module Danger
           end
 
           matching_comments = danger_comments.select do |comment_data|
-            if comment_data["path"] == m.file && comment_data["position"] == position
-              # Parse it to avoid problems with strikethrough
-              violation = violations_from_table(comment_data["body"]).first
-              if violation
-                messages_are_equivalent(violation, m)
-              else
-                blob_regexp = %r{blob/[0-9a-z]+/}
-                comment_data["body"].sub(blob_regexp, "") == body.sub(blob_regexp, "")
-              end
-            else
+            position = comment_data["position"]
+
+            if position.nil?
               false
+            else
+              position["new_path"] == m.file && position["new_line"] == m.line
             end
           end
 
@@ -362,7 +399,7 @@ module Danger
 
             # Update the comment to remove the strikethrough if present
             comment = matching_comments.first
-            client.update_merge_request_comment(ci_source.repo_slug, comment["id"], body)
+            client.update_merge_request_discussion_note(ci_source.repo_slug, ci_source.pull_request_id, comment["discussion_id"], comment["id"], body)
           end
 
           # Remove this element from the array
